requirements.txt
flask # serving dynamic pages
requests # HTTP calls to API
pydantic # structured response
gunicorn  # deploying with Nginx
python-dotenv  # For managing API keys (if needed)
loguru  # For better logging